@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import multiprocessing
 import sys
 import threading
@@ -69,24 +71,28 @@ def test_cross_module_gil_inner_pybind11_acquired():
     m.test_cross_module_gil_inner_pybind11_acquired()
 
 
+@pytest.mark.skipif(sys.platform.startswith("emscripten"), reason="Requires threads")
 def test_cross_module_gil_nested_custom_released():
     """Makes sure that the GIL can be nested acquired/released by another module
     from a GIL-released state using custom locking logic."""
     m.test_cross_module_gil_nested_custom_released()
 
 
+@pytest.mark.skipif(sys.platform.startswith("emscripten"), reason="Requires threads")
 def test_cross_module_gil_nested_custom_acquired():
     """Makes sure that the GIL can be nested acquired/acquired by another module
     from a GIL-acquired state using custom locking logic."""
     m.test_cross_module_gil_nested_custom_acquired()
 
 
+@pytest.mark.skipif(sys.platform.startswith("emscripten"), reason="Requires threads")
 def test_cross_module_gil_nested_pybind11_released():
     """Makes sure that the GIL can be nested acquired/released by another module
     from a GIL-released state using pybind11 locking logic."""
     m.test_cross_module_gil_nested_pybind11_released()
 
 
+@pytest.mark.skipif(sys.platform.startswith("emscripten"), reason="Requires threads")
 def test_cross_module_gil_nested_pybind11_acquired():
     """Makes sure that the GIL can be nested acquired/acquired by another module
     from a GIL-acquired state using pybind11 locking logic."""
@@ -101,6 +107,7 @@ def test_nested_acquire():
     assert m.test_nested_acquire(0xAB) == "171"
 
 
+@pytest.mark.skipif(sys.platform.startswith("emscripten"), reason="Requires threads")
 def test_multi_acquire_release_cross_module():
     for bits in range(16 * 8):
         internals_ids = m.test_multi_acquire_release_cross_module(bits)
@@ -202,7 +209,7 @@ def _run_in_threads(test_fn, num_threads, parallel):
         thread.join()
 
 
-# TODO: FIXME, sometimes returns -11 (segfault) instead of 0 on macOS Python 3.9
+@pytest.mark.skipif(sys.platform.startswith("emscripten"), reason="Requires threads")
 @pytest.mark.parametrize("test_fn", ALL_BASIC_TESTS_PLUS_INTENTIONAL_DEADLOCK)
 def test_run_in_process_one_thread(test_fn):
     """Makes sure there is no GIL deadlock when running in a thread.
@@ -212,7 +219,7 @@ def test_run_in_process_one_thread(test_fn):
     assert _run_in_process(_run_in_threads, test_fn, num_threads=1, parallel=False) == 0
 
 
-# TODO: FIXME on macOS Python 3.9
+@pytest.mark.skipif(sys.platform.startswith("emscripten"), reason="Requires threads")
 @pytest.mark.parametrize("test_fn", ALL_BASIC_TESTS_PLUS_INTENTIONAL_DEADLOCK)
 def test_run_in_process_multiple_threads_parallel(test_fn):
     """Makes sure there is no GIL deadlock when running in a thread multiple times in parallel.
@@ -222,7 +229,7 @@ def test_run_in_process_multiple_threads_parallel(test_fn):
     assert _run_in_process(_run_in_threads, test_fn, num_threads=8, parallel=True) == 0
 
 
-# TODO: FIXME on macOS Python 3.9
+@pytest.mark.skipif(sys.platform.startswith("emscripten"), reason="Requires threads")
 @pytest.mark.parametrize("test_fn", ALL_BASIC_TESTS_PLUS_INTENTIONAL_DEADLOCK)
 def test_run_in_process_multiple_threads_sequential(test_fn):
     """Makes sure there is no GIL deadlock when running in a thread multiple times sequentially.
@@ -232,7 +239,7 @@ def test_run_in_process_multiple_threads_sequential(test_fn):
     assert _run_in_process(_run_in_threads, test_fn, num_threads=8, parallel=False) == 0
 
 
-# TODO: FIXME on macOS Python 3.9
+@pytest.mark.skipif(sys.platform.startswith("emscripten"), reason="Requires threads")
 @pytest.mark.parametrize("test_fn", ALL_BASIC_TESTS_PLUS_INTENTIONAL_DEADLOCK)
 def test_run_in_process_direct(test_fn):
     """Makes sure there is no GIL deadlock when using processes.
